+++ Braintree.Tests/QueryStringTest.cs
@@ -67,7 +67,7 @@ namespace Braintree.Tests
             };
 
             String actual = new QueryString().Append("[credit_card]", request).ToString();
-            Assert.AreEqual("%5bcredit_card%5d%5bcardholder_name%5d=Drew&%5bcredit_card%5d%5bcvv%5d=123&%5bcredit_card%5d%5boptions%5d%5bverify_card%5d=true&%5bcredit_card%5d%5boptions%5d%5bmake_default%5d=true", actual);
+            Assert.AreEqual("%5bcredit_card%5d%5bcardholder_name%5d=Drew&%5bcredit_card%5d%5bcvv%5d=123&%5bcredit_card%5d%5boptions%5d%5bmake_default%5d=true&%5bcredit_card%5d%5boptions%5d%5bverify_card%5d=true", actual);
         }
     }
 }