+++ Braintree/CreditCardVerification.cs
@@ -13,6 +13,7 @@ namespace Braintree
         public String AvsPostalCodeResponseCode { get; protected set; }
         public String AvsStreetAddressResponseCode { get; protected set; }
         public String CvvResponseCode { get; protected set; }
+        public TransactionGatewayRejectionReason GatewayRejectionReason { get; protected set; }
         public String ProcessorResponseCode { get; protected set; }
         public String ProcessorResponseText { get; protected set; }
         public String MerchantAccountId { get; protected set; }
@@ -26,6 +27,11 @@ namespace Braintree
             AvsPostalCodeResponseCode = node.GetString("avs-postal-code-response-code");
             AvsStreetAddressResponseCode = node.GetString("avs-street-address-response-code");
             CvvResponseCode = node.GetString("cvv-response-code");
+            GatewayRejectionReason = (TransactionGatewayRejectionReason)CollectionUtil.Find(
+                TransactionGatewayRejectionReason.ALL,
+                node.GetString("gateway-rejection-reason"),
+                null
+            );
             ProcessorResponseCode = node.GetString("processor-response-code");
             ProcessorResponseText = node.GetString("processor-response-text");
             MerchantAccountId = node.GetString("merchant-account-id");