+++ Braintree/Transaction.cs
@@ -7,6 +7,20 @@ using System.Xml;
 
 namespace Braintree
 {
+    public class TransactionGatewayRejectionReason : Enumeration
+    {
+        public static readonly TransactionGatewayRejectionReason AVS = new TransactionGatewayRejectionReason("avs");
+        public static readonly TransactionGatewayRejectionReason AVS_AND_CVV = new TransactionGatewayRejectionReason("avs_and_cvv");
+        public static readonly TransactionGatewayRejectionReason CVV = new TransactionGatewayRejectionReason("cvv");
+        public static readonly TransactionGatewayRejectionReason DUPLICATE = new TransactionGatewayRejectionReason("duplicate");
+
+        public static readonly TransactionGatewayRejectionReason[] ALL = {
+            AVS, AVS_AND_CVV, CVV, DUPLICATE
+        };
+
+        protected TransactionGatewayRejectionReason(String name) : base(name) {}
+    }
+
     public class TransactionStatus : Enumeration
     {
         public static readonly TransactionStatus AUTHORIZED = new TransactionStatus("authorized");
@@ -17,13 +31,12 @@ namespace Braintree
         public static readonly TransactionStatus SETTLED = new TransactionStatus("settled");
         public static readonly TransactionStatus SETTLEMENT_FAILED = new TransactionStatus("settlement_failed");
         public static readonly TransactionStatus SUBMITTED_FOR_SETTLEMENT = new TransactionStatus("submitted_for_settlement");
-        public static readonly TransactionStatus UNKNOWN = new TransactionStatus("unknown");
         public static readonly TransactionStatus UNRECOGNIZED = new TransactionStatus("unrecognized");
         public static readonly TransactionStatus VOIDED = new TransactionStatus("voided");
 
         public static readonly TransactionStatus[] ALL = {
             AUTHORIZED, AUTHORIZING, FAILED, GATEWAY_REJECTED, PROCESSOR_DECLINED, SETTLED,
-            SETTLEMENT_FAILED, SUBMITTED_FOR_SETTLEMENT, UNKNOWN, VOIDED
+            SETTLEMENT_FAILED, SUBMITTED_FOR_SETTLEMENT, VOIDED
         };
 
         protected TransactionStatus(String name) : base(name) {}
@@ -83,6 +96,7 @@ namespace Braintree
         public String CurrencyIsoCode { get; protected set; }
         public Customer Customer { get; protected set; }
         public String CvvResponseCode { get; protected set; }
+        public TransactionGatewayRejectionReason GatewayRejectionReason { get; protected set; }
         public String MerchantAccountId { get; protected set; }
         public String OrderId { get; protected set; }
         public String ProcessorAuthorizationCode { get; protected set; }
@@ -107,6 +121,11 @@ namespace Braintree
             AvsErrorResponseCode = node.GetString("avs-error-response-code");
             AvsPostalCodeResponseCode = node.GetString("avs-postal-code-response-code");
             AvsStreetAddressResponseCode = node.GetString("avs-street-address-response-code");
+            GatewayRejectionReason = (TransactionGatewayRejectionReason)CollectionUtil.Find(
+                TransactionGatewayRejectionReason.ALL,
+                node.GetString("gateway-rejection-reason"),
+                null
+            );
             OrderId = node.GetString("order-id");
             Status = (TransactionStatus)CollectionUtil.Find(TransactionStatus.ALL, node.GetString("status"), TransactionStatus.UNRECOGNIZED);